+++ test/integration/resultsPage.js
@@ -30,14 +30,6 @@ function setTypeAndOriginPairs() {
       origin: constants.serviceChoices.over25,
       type: constants.serviceTypes.professional,
     },
-    {
-      origin: constants.serviceChoices['16to24'],
-      type: constants.serviceTypes.kit,
-    },
-    {
-      origin: constants.serviceChoices.over25,
-      type: constants.serviceTypes.kit,
-    },
   ];
 }